drm: linlondp: merge updates from 26q2
This commit cherry-picks the following commits from 26q2:

- VO: DPTSW-22606: Stop KMS poll before DP PM teardown
- DPTSW-23684: use aperture_remove_all_conflicting_devices to disable uefi fb
- DPTSW-24293: drm/cix: linlon display cluster bring-up
- DPTSW-23794: drm/cix: dpu pipeline binding and source_id consolidation
- DPTSW-24293: drm/cix: ACPI linlon cluster and DPTx possible_crtcs
- DPTSW-24293: drm/cix: single master drm, drop component bind bypass
- DPTSW-23794: dptx: derive SST source_id from peer pipeline graph
- DPTSW-24781: drm/cix: route direct shutdown through PM suspend
- DPTSW-24785: drm/cix: linlon-dp: cluster PM for modeset suspend/resume
- DPTSW-24834: use aperture_remove_all_conflicting_devices to disable uefi fb
- DPTSW-24753: linlon-dp: map cluster IOMMU on demand
- DPTSW-24897: linlon: fix cluster DPU GenPD runtime PM ordering
- DPTSW-24897: linlon: cluster slave runtime PM after linlondp_dev_create
- DPTSW-19407 : screen flicker issue during reboot
- VO: DPTSW-23827: trilin_dptx avoid NULL device_link_del on unbind
- DPTSW-24977: dptx: sync SST dp_panel.source_id from Linlon CRTC master pipeline
- DPTSW-24990: linlon-dp: skip missing cluster DPU platform devices directly
- DPTSW-25100: verify sink link before GOP fast path
- DPTSW-25168: fix str common issue for display
- DPTSW-25240: DRM/cix: dptx: fix GOP HPD notify on first boot
- DPTSW-25645: fix Redmi-L50RB-RAE TV display issue
- DPTSW-25678: drm/cix: linlon-dp: fix CP6187 ACPI cluster DPU probe crash
- DPTSW-22258: DRM/cix: linlon-dp STR EDID fingerprint fast resume
- DPTSW-22258: DRM/cix: linlon-dp: fix STR resume after blank-before-suspend
- DPTSW-22258: drm/cix: trilin-dptx: gate eDP panel ops on DPCD and HPD
- DPTSW-22808: DRM/cix: dptx: hold off PSR entry 1s after full enable
- DPTSW-25854: DRM/cix: dptx: skip PSR resync wait on full disable
- DPTSW-26233: linlon-dp: drop stale pre-blank when no display connected
- DPTPLAT-723: drm/cix: fix device link stateless flag and eDP panel probe deferral issues
- DPTSW-26233: linlon-dp: drop stale pre-blank when no display connected When linlondp_cluster_pm_resume, there is no display connected, linlondp_pm_conn_fp_match() report dismatched directly.

+++ drivers/gpu/drm/arm/display/komeda/komeda_kms.h
@@ -77,7 +77,7 @@ struct komeda_crtc {
 	struct komeda_pipeline *slave;
 
 	/** @slave_planes: komeda slave planes mask */
-	u32 slave_planes;
+	u64 slave_planes;
 
 	/** @wb_conn: komeda write back connector */
 	struct komeda_wb_connector *wb_conn;
